refactor(frontend): extract actions menu + file-attach into actions.ts (#312)
Move the mobile actions (+) menu and the file-attachment flow (#286) out of
main.ts into actions.ts (~184 lines). Clean extraction: no state
reassignment, deps are getActiveTerminal/showToast/activeSessionId/sessions
(main), runPasteFlow (paste.ts), uploadSessionFiles (api). DOM re-queried
locally; the two shared consts (actionsBtn/actionsMenu, used by the chrome
toggle + Escape guard) stay in main.ts; closeActionsMenu is exported for its
one main.ts caller. Same TDZ-safe pattern.

main.ts: 1042 → 876. tsc + Biome clean, vite build green, tests pass (66).
